requirements.txt
langchain-core~=0.3.58
langgraph~=0.4.8
langchain-community~=0.3.12
langchain-google-genai~=2.1.5
langchain-tavily~=0.2.5
python-dotenv~=1.1.0
pydantic~=2.10.3